refactor(tests): split test_route_detector.py (1516→5 files, phase8 slice)
Split 22 test classes across 5 focused modules (all under 500 lines):
- test_route_detector_python.py — Flask/FastAPI/K1 framework detection
- test_route_detector_web.py — Express/Spring detection
- test_route_detector_go.py — Go frameworks + summary/dispatch/source-walk
- test_route_detector_tool.py — tool schema and execute
- test_route_detector_cache.py — cache persistence, handler quality, envelope

Shared fixtures (flask_project, fastapi_project, express_project,
spring_project, go_*, multi_framework_project) extracted to a new
tests/unit/conftest.py auto-discovered by pytest.

86 tests pass before and after; full unit suite (16 598 tests) green.
